+++ tests/testthat/test-rsqlserver.R
@@ -12,14 +12,15 @@ test_that("get_server_details works", {
   expect_error(get_server_details("SQL_PORT_ERROR", file))
 })
 
-test_that("jtds_url works", {
+test_that("msft_url works", {
   file <- system.file("extdata", "sql.yaml", package = "RSQLServer")
   sd <- get_server_details("SQL_PROD", file)
-  expected_url <- "jdbc:jtds:sqlserver://11.1.111.11:1433/test"
-  expect_equal(jtds_url(sd$server, sd$type, sd$port, "test"), expected_url)
-  expected_url <- "jdbc:jtds:sqlserver://11.1.111.11:1433/test;useNTLMv2=true;domain=corpname"
-  expect_equal(jtds_url(sd$server, sd$type, sd$port, "test",
-    list(useNTLMv2 = 'true', domain = 'corpname')), expected_url)
+  expected_url <- "jdbc:sqlserver://;serverName=11.1.111.11;portNumber=1433;database=test"
+  expect_equal(msft_url(sd$server, sd$port, sd$instance, list(database = "test")),
+    expected_url)
+  expected_url <- "jdbc:sqlserver://;serverName=11.1.111.11;portNumber=1433;authentication=ActiveDirectoryIntegrated;domain=corpname;database=test"
+  expect_equal(msft_url(sd$server, sd$port, sd$instance,
+    list(authentication = 'ActiveDirectoryIntegrated', domain = 'corpname', database = "test")), expected_url)
 })
 
 test_that("have_test_server and SQLServerConnection work", {