+++ NEWS.md
@@ -1,3 +1,9 @@
+# Version 0.3.099
+
+We have replaced the jTDS drivers with Microsoft's SQL Server JDBC driver. This has resulted in a number of interface changes:
+
+* The `type` argument to `dbConnect()` is unsupported and has been soft-deprecated. This will become defunct at the next release.
+
 # Version 0.2.099
 
 RSQLServer was archived by CRAN after dplyr v0.4 irredeemably broke the dplyr SQL Server backend provided by this package. Well we are back on CRAN and there have been a lot of changes since you last saw this package including improvements to the DBI backend, compatibility with the latest iteration of the dplyr/dbplyr backend and removing the reliance on RJDBC's middleware.